requirements.txt
asgiref==3.8.1
attrs==24.3.0
bidict==0.23.1
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.28.0
eventlet==0.38.2
greenlet==3.1.1
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2==2.9.10
PyJWT==2.10.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
simple-websocket==1.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
wsproto==1.2.0
